# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kids Art Tales API
  description: API for children's drawing app with image-to-story, image-to-cartoon, and image-to-animation features
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/media/image-to-voice:
    post:
      summary: Generate audio story from image and voice keywords
      description: Analyzes image and combines with user voice keywords to generate a story with audio
      tags:
        - Media
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: User's drawing image
                voiceText:
                  type: string
                  description: Keywords from user voice input
                userId:
                  type: string
                  description: User ID
      responses:
        '200':
          description: Story and audio generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  keywords:
                    type: array
                    items:
                      type: string
                  story:
                    type: string
                  audioUrl:
                    type: string
                  userId:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
  /api/media/image-to-image:
    post:
      summary: Generate cartoon image from user's drawing
      description: Converts user's drawing to cartoon style based on keywords
      tags:
        - Media
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: User's drawing image
                prompt:
                  type: string
                  description: Keywords for cartoon generation
                userId:
                  type: string
                originalImageUrl:
                  type: string
      responses:
        '200':
          description: Cartoon image generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  prompt:
                    type: string
                  originalImageUrl:
                    type: string
                  cartoonImageUrl:
                    type: string
                  userId:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
  /api/media/image-to-video:
    post:
      summary: Generate animation video from image
      description: Creates a short animation (under 1 minute) from user's drawing
      tags:
        - Media
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                prompt:
                  type: string
                userId:
                  type: string
      responses:
        '200':
          description: Animation video generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  prompt:
                    type: string
                  videoUrl:
                    type: string
                  durationSec:
                    type: integer
                  userId:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
  /api/media/generate-picture-book:
    post:
      summary: Generate picture book from cartoon image
      description: Creates a 4-6 page picture book story from cartoon image and user interaction
      tags:
        - Media
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cartoonImage:
                  type: string
                  format: binary
                cartoonImageUrl:
                  type: string
                prompt:
                  type: string
                userId:
                  type: string
                pageCount:
                  type: integer
                  minimum: 4
                  maximum: 6
                  default: 4
      responses:
        '200':
          description: Picture book generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  pages:
                    type: array
                    items:
                      type: object
                      properties:
                        pageNumber:
                          type: integer
                        imageUrl:
                          type: string
                        text:
                          type: string
                  pageCount:
                    type: integer
                  coverImageUrl:
                    type: string
                  userId:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
  /api/media/cocreate:
    post:
      summary: Co-create content from multiple works
      description: Generate new content by combining multiple user works
      tags:
        - Media
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workIds
                - type
                - userId
              properties:
                workIds:
                  type: array
                  items:
                    type: string
                  minItems: 2
                type:
                  type: string
                  enum: [picture-book, animation, audio-story]
                userId:
                  type: string
                prompt:
                  type: string
      responses:
        '200':
          description: Co-created content generated
        '400':
          description: Invalid request (need at least 2 work IDs)

  # User Management APIs
...